pyproject.toml
[tool.poetry]
name = "sound_cloud"
version = "0.1.0"
description = ""
authors = ["Aleksan <Dev.aleksan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
environs = "^9.3.4"
psycopg2 = "^2.9.1"
Django = "^3.2.9"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.10.0"
drf-yasg = "^1.20.0"
Pillow = "^8.4.0"
requests = "^2.26.0"
google-auth = "^2.3.3"
PyJWT = "^2.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

